Refactor: Move partition layout logic into cos.go
* Upgrade yip to 0.9.25 so we could use latest yip/schema package to
  define the "layout" stage.

* Move function `console.createPartitionLayout` to
  `config.createPartitioningStage` in cos.go as this make much more
  sense.
